requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
mysql-connector-python==8.0.26
protobuf==3.19.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
stripe==2.63.0
toml==0.10.2
tzdata==2021.5
urllib3==1.26.7

aioredis==1.3.1
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
dj-database-url==0.5.0
django-csp==3.7
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
incremental==17.5.0
msgpack==1.0.0
psycopg2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
service-identity==18.1.0
six==1.15.0
Twisted
txaio==20.4.1
zope.interface==5.1.0
